save new xml
Former-commit-id: b75a22207664c4a1d69702d46355d23af2ee84a7

+++ trunk/GeoImageViewer/src/main/java/org/geoimage/viewer/core/batch/Sumo.java
@@ -135,7 +135,7 @@ public class Sumo {
 	        if(status!=PARAM_ERROR){
 	        	int index=inputParams.indexOf(BUFFER_PARAM);
 	        	if(index!=-1){
-	        		params.buffer=Double.parseDouble(inputParams.get(index+1));
+	        		params.buffer=Integer.parseInt(inputParams.get(index+1));
 	        	}
 	
 	        	//set the treshold params
